# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/movie": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set watchlist or watched status for a movie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Upsert movie state",
                "parameters": [
                    {
                        "description": "Movie state",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_movie.UpsertStateRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movie/discover": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Browse the local movie cache with TMDB-style filters and sort. Each result is enriched with full TMDB detail via append_to_response.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Discover movies",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "popularity.desc | release_date.desc | vote_average.desc | title.asc | watched_at.desc | watchlisted_at.desc | ...",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated genre IDs (AND logic)",
                        "name": "with_genres",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated genre IDs to exclude",
                        "name": "without_genres",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by release year",
                        "name": "primary_release_year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Release date \u003e= (YYYY-MM-DD)",
                        "name": "release_date.gte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Release date \u003c= (YYYY-MM-DD)",
                        "name": "release_date.lte",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Vote average \u003e=",
                        "name": "vote_average.gte",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Vote average \u003c=",
                        "name": "vote_average.lte",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Vote count \u003e=",
                        "name": "vote_count.gte",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO 639-1 language code (e.g. en, th)",
                        "name": "with_original_language",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include adult content (default false)",
                        "name": "include_adult",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by softcore flag",
                        "name": "softcore",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated provider IDs",
                        "name": "with_watch_providers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ISO 3166-1 country code for watch provider filter",
                        "name": "watch_region",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "watchlist | watched",
                        "name": "with_account_status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movie/random": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns movies the user has no watchlist/watched record for, preferring upcoming releases and falling back to popular titles. Results are randomized on every call.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Random movie suggestions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (echoed back, does not affect sampling; default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated account statuses to exclude, e.g. watchlist,watched",
                        "name": "without_status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results (default 20)",
                        "name": "total",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movie/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Search TMDB for movies. Results are enriched with genres and metadata from the local cache when available.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Search movies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movie/states": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all movie watchlist/history records for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "List movie states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieUser"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/movie/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns full TMDB detail for a movie, with vote_average/vote_count from the local cache.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "movies"
                ],
                "summary": "Get movie detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "TMDB movie id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_movie.MovieResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.MovieResponse"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
        "github_com_peera_movizius-go-service_internal_shared_response.Page-internal_movie_MovieUser": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.MovieUser"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_results": {
                    "type": "integer"
                }
            }
        },
...
        "internal_movie.Casts": {
            "type": "object",
            "properties": {
                "cast": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.Cast"
                    }
                },
                "crew": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.Cast"
                    }
                }
            }
        },
        "internal_movie.Collection": {
            "type": "object",
            "properties": {
                "backdrop_path": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "poster_path": {
                    "type": "string"
                }
            }
        },
        "internal_movie.ExternalIDs": {
            "type": "object",
            "properties": {
                "facebook_id": {
                    "type": "string"
                },
                "imdb_id": {
                    "type": "string"
                },
                "instagram_id": {
                    "type": "string"
                },
                "twitter_id": {
                    "type": "string"
                },
                "wikidata_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_movie.Genre": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_movie.MovieResponse": {
            "type": "object",
            "properties": {
                "adult": {
                    "type": "boolean"
                },
                "backdrop_path": {
                    "type": "string"
                },
                "belongs_to_collection": {
                    "$ref": "#/definitions/internal_movie.Collection"
                },
                "budget": {
                    "type": "integer"
                },
                "casts": {
                    "$ref": "#/definitions/internal_movie.Casts"
                },
                "external_ids": {
                    "$ref": "#/definitions/internal_movie.ExternalIDs"
                },
                "genres": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.Genre"
                    }
                },
                "homepage": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "imdb_id": {
                    "type": "string"
                },
                "media_type": {
                    "type": "string"
                },
                "origin_country": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "original_language": {
                    "type": "string"
                },
                "original_title": {
                    "type": "string"
                },
                "overview": {
                    "type": "string"
                },
                "popularity": {
                    "type": "number"
                },
                "poster_path": {
                    "type": "string"
                },
                "production_companies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.ProductionCompany"
                    }
                },
                "production_countries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.ProductionCountry"
                    }
                },
                "release_date": {
                    "type": "string"
                },
                "release_date_th": {
                    "type": "string"
                },
                "release_dates": {
                    "$ref": "#/definitions/internal_movie.ReleaseDates"
                },
                "revenue": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "integer"
                },
                "softcore": {
                    "type": "boolean"
                },
                "spoken_languages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_movie.SpokenLanguage"
                    }
                },
                "status": {
                    "type": "string"
                },
                "tagline": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "boolean"
                },
                "videos": {
                    "$ref": "#/definitions/internal_movie.Videos"
                },
                "vote_average": {
                    "type": "number"
                },
                "vote_count": {
                    "type": "integer"
                },
                "watch_providers": {
                    "$ref": "#/definitions/internal_movie.WatchProviderCountry"
                }
            }
        },
        "internal_movie.MovieUser": {
            "type": "object",
            "properties": {
                "account_status": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "media_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "watched_at": {
                    "type": "string"
                },
                "watchlisted_at": {
                    "type": "string"
                }
            }
        },
...
        "internal_movie.ProductionCountry": {
            "type": "object",
            "properties": {
                "iso_3166_1": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "internal_movie.UpsertStateRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}